fix: battery JSON, sleep timeout boot race, brightness pending flag
- int8_t bat treated as char by Arduino String concat, corrupting
  the /api/status JSON and causing the battery pill to never update;
  cast to int to get decimal representation
- lastActivityMs_ initialized to 0, so millis() at loop() start
  (post-WiFi/NTP, ~10-25s) immediately exceeded short timeouts;
  set to millis() in begin() so timeout counts from ready state
- brightness change via API now sets pendingBrightness_ flag and
  lets update() apply it in display loop context rather than web
  handler context

+++ espibright/WebAPI.cpp
@@ -125,7 +125,7 @@ void WebAPI::handleApiPackets_() {
 
 void WebAPI::handleApiStatus_() {
     sendCors_();
-    int8_t bat = M5.Power.getBatteryLevel();
+    int bat = M5.Power.getBatteryLevel();
     String j = "{\"label\":\""  + String(rf_.lastLabel) + "\""
              + ",\"hex\":\""    + String(rf_.lastHex)   + "\""
              + ",\"ms_ago\":"   + (rf_.lastMs ? String(millis() - rf_.lastMs) : String("null"))
